Reorganize Yaml serialization code
No refactoring going on here, just moving all Yaml serialization code
to the `SAGESharp.IO.Yaml` namespace and split the *Yaml.cs* file
into individual files for each class.
