+++ SAGESharp.Tests/IO/Yaml/SLBTypeInspectorTests.cs
@@ -9,12 +9,13 @@ using NSubstitute.ClearExtensions;
 using NUnit.Framework;
 using NUtils.Extensions;
 using SAGESharp.IO;
+using SAGESharp.IO.Yaml;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using YamlDotNet.Serialization;
 
-namespace SAGESharp.Tests.IO
+namespace SAGESharp.Tests.IO.Yaml
 {
     class SLBTypeInspectorTests
     {
